Fix z icon.
