+++ src/UniversalDashboard.Materialize/Components/ud-icon.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default class UdIcon extends React.Component {
+    render() {
+        return UniversalDashboard.renderComponent({
+            ...this.props,
+            type: 'icon'
+        });
+    }
+}
